Support JWTs with authority claims are a concatenated string
When HMPPS Auth creates a certain type of JWT it will add authorities as a string[1]. Given this is possible (and what CAS2 are using) we need to be able to unpack them safely.

We’ve had to do similar in our API [2]

Here’s an example of a simpler JWT where authorities come out as a string:

```
{
  "jti": "11d0f107-c0f9-4896-aa4f-71746cb0b046",
  "sub": "POM_USER",
  "authorities": "ROLE_POM,ROLE_PRISON",
  "name": "Pom User",
  "auth_source": "nomis",
  "user_id": "29",
  "passed_mfa": false,
  "exp": 1713377737
}
```

This change ensures that authorities can be cast correctly and we don’t get a 500.

[1] https://github.com/ministryofjustice/hmpps-auth/blob/9b6b16a5cfc3604369280c26dd14ece99749bac0/src/main/kotlin/uk/gov/justice/digital/hmpps/oauth2server/security/JwtAuthenticationHelper.kt#L86
[2] https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/e4a37c52d7a25715801f803186f5ca6798d88b0f/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/config/OAuth2ResourceServerSecurityConfiguration.kt#L137